Move general group to separate class
